does data binding, creates movie controller, does refactoring of some parts of the code, adds some small changes
