+++ src/components/amount-of-movies.js
@@ -1,4 +1,4 @@
-import AbstractComponent from "./abstractComponent.js";
+import AbstractComponent from "./abstract-component.js";
 
 export const createAmountOfMovies = (amount) => (`<p>${amount} movies inside</p>`);
