+++ src/components/btn-show-more.js
@@ -1,4 +1,4 @@
-import AbstractComponent from "./abstractComponent.js";
+import AbstractComponent from "./abstract-component.js";
 
 const createBtnShowMore = () => (`<button class="films-list__show-more">Show more</button>`);
